Breakthrough time comparison for homo and heterogeneous matrix porosity model.
This project was done to obtain a working mathematical model for 1-D two-phase flow compressible flow which
can be used to estimate the effects of water-flooding on a dual-porosity system. It also goes through the analysis of
saturation and pressure graphs to estimate the approximate breakthrough time for the given physical parameters
such as constant injection rate and constant outlet pressure.
In the end, comparative analysis in using a homogeneous and heterogeneous matrix porosity model was performed.